+++ docs/tr2/CHANGELOG.md
@@ -19,9 +19,11 @@
 - added an option to display level counter in the statistics dialog (Graphic options → UI → Level counter) (#1087)
 - added an option to control playing of certain animation sound effects such as doors when underwater (Sound options → Underwater animation SFX) (#3385)
 - added an option to choose between original TR1, original TR2 or unrestricted look modes (Gameplay settings → Controls → Look mode) (#3403)
+- added an option to set the upscaling filter (Graqphic settings → Rendering → Upscaling filter)
 - added an inverted look camera option (Gameplay settings → Controls → Inverted look) (#3403)
 - changed statistics details mode to be placed in the UI section
 - added missing end of level statistic screens to Home Sweet Home and Kingdom (#2682)
+- changed the "sizer" option name to "Upscaling factor" (Graphic settings → Rendering → Upscaling factor)
 - changed controls dialog to remember the player's preferred input method
 - changed UI to show icons relevant to the chosen input method
 - changed death timer skip to only trigger with Action and Inventory keys